+++ .env
@@ -1,1 +1,7 @@
 VITE_BASE_URL = 'https://rolling-api.vercel.app'
+
+#첫번째 offset (prev 버튼 X)
+VITE_POPULAR_FIRST_URL="https://rolling-api.vercel.app/18-2/recipients/?limit=4&&sort=like"
+
+#첫번째 offset (prev 버튼 X)
+VITE_NEW_FIRST_URL="https://rolling-api.vercel.app/18-2/recipients/?limit=4&"
